Make the cleanup gated behind a flag

+++ TeamCode/src/main/scripts/autoauto-compiler/compiler/transmutations/codebasewide-tasks/postprocess/zz-cleanup/index.js
@@ -1,5 +1,6 @@
 "use strict";
 
+const commandLineInterface = require("../../../../../../command-line-interface");
 const safeFsUtils = require("../../../../../../script-helpers/safe-fs-utils");
 const deleteOldGeneratedFiles = require("./delete-old-generated-files");
 
@@ -7,16 +8,19 @@ const deleteOldGeneratedFiles = require("./delete-old-generated-files");
  * @type {import("../../..").CodebaseTransmutateFunction}
  */
 module.exports = function (context, contexts) {
-    var newFiles = Object.keys(context.writtenFiles);
     
-    for(const ctx of contexts) {
-        if(ctx.success === "SUCCESS") {
-            newFiles.push(...Object.keys(ctx.fileContext.writtenFiles));
+    if(commandLineInterface["run-cleanup"] === true) {
+        var newFiles = Object.keys(context.writtenFiles);
+        
+        for(const ctx of contexts) {
+            if(ctx.success === "SUCCESS") {
+                newFiles.push(...Object.keys(ctx.fileContext.writtenFiles));
+            }
         }
-    }
-    
+        
 
-    safeFsUtils.cleanDirectory(context.resultRoot, newFiles);
-    
-    deleteOldGeneratedFiles();
+        safeFsUtils.cleanDirectory(context.resultRoot, newFiles);
+        
+        deleteOldGeneratedFiles();
+    }
 }
